backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
motor>=3.3.2
pydantic>=2.5.0
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
email-validator>=2.1.0
fastapi-crudrouter>=0.8.6
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
aiofiles>=23.2.1
psutil>=5.9.0
